requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
email_validator==2.1.1
Jinja2==3.1.3
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
aiohttp==3.9.5
asyncpg==0.29.0
fastapi-users-db-sqlalchemy==6.0.1
pytest==8.2.2
